chore: Replace artwork favoriting logic
